requirements.txt
allure-pytest==2.13.2
pytest==7.4.3
requests==2.31.0
jsondiff==2.0.0
